requirements.txt
gradio==4.36.1
langchain==0.2.5
langchain-openai==0.1.8
openai==1.34.0
python-dotenv==1.0.1
